second commit

+++ src/app/categories/categories/categories.component.ts
@@ -2,20 +2,22 @@ import { CategoriesService } from './../categories.service';
 import { CategoryDetailsComponent } from './../category-details/category-details.component';
 import { Category } from './../../data models/category';
 import { Component, OnInit } from '@angular/core';
-import { DialogService } from 'primeng/api';
+import { DialogService, ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'app-categories',
   templateUrl: './categories.component.html',
   styleUrls: ['./categories.component.css'],
-  providers: [DialogService]
+  providers: [DialogService, ConfirmationService]
 })
 export class CategoriesComponent implements OnInit {
 
   selectdCategory: Category = null;
   catergories: Category[];
 
-  constructor(public dialogService: DialogService, private cs: CategoriesService) { }
+  constructor(public dialogService: DialogService,
+              private cs: CategoriesService,
+              private confirmationService: ConfirmationService) { }
 
   ngOnInit() {
     this.cs.categories.subscribe(categories => this.catergories = categories);
@@ -26,7 +28,7 @@ export class CategoriesComponent implements OnInit {
       header: header,
       width: '500px',
       data: {
-        name : this.selectdCategory? this.selectdCategory.name : '',
+        category : this.selectdCategory,
         edit: edit
       } 
     });
@@ -62,8 +64,13 @@ export class CategoriesComponent implements OnInit {
         this.show(`Category ${this.selectdCategory.name} details`, false);
         break;
       case 'delete':
-        this.cs.deleteCategory(this.selectdCategory.id);
-        this.selectdCategory = null;
+        this.confirmationService.confirm({
+            message: "Are you sure you want to delete category and all it's loctions?",
+            accept: () => {
+              this.cs.deleteCategory(this.selectdCategory.id);
+              this.selectdCategory = null;
+            }
+        });
         break;
     }
   }
